pyproject.toml
[tool.poetry]
name = "fluxvault"
version = "0.6.0"
description = "A system to load secrets into Flux applications"
authors = ["Tom Moulton <tom@moulton.us>", "<David White <dr.white.nz@gmail.com>"]
license = "MIT"
repository = "https://github.com/RunOnFlux/FluxVault.git"
homepage = "https://runonflux.io"

[tool.poetry.dependencies]
python = "^3.8"
aiotinyrpc = {version = "^0.8.3", extras = ["socket"]}
requests = "^2.28.1"
aiofiles = "^22.1.0"
aioshutil = "^1.2"
aiohttp = "^3.8.3"
python-daemon = "^2.3.2"
typer = "^0.7.0"
dnspython = "^2.2.1"
rich = "^12.6.0"
keyring = "^23.11.0"
ownca = "^0.3.3"
randomname = "^0.1.5"
python-socketio= "^5.7.2"

[tool.isort]
profile = "black"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
fluxvault = "fluxvault.cli:entrypoint"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
